Fix for Phantom Canyon (NUC11) hardware
On a NUC11PHBi7 under CentOS 7, all of the driver's WMI calls failed
with errors like:

    ACPI Error: Field [CMD4] at 40 exceeds Buffer [NULL] size 32 (bits) (20130517/dsopcode-236)
    ACPI Error: Method parse/execution failed [\AMW0.WMAA] (Node ffff8f9bf62a08d0), AE_AML_BUFFER_LIMIT (20130517/psparse-536)
    ACPI Exception: AE_AML_BUFFER_LIMIT, wmi_evaluate_method (20130517/NUC_LED-266)

Simply increasing the size of the input buffer passed to
wmi_evaluate_method by one byte silences those errors, and allows the
driver to work correctly on this hardware.

On a NUC8i7HVK, the additional byte makes no difference to the
operation of the driver.